This is the version which knows about o multiple rules for a target o multiple targets in a rule but does not allow more than one command for a target. It runs fine through all its tests and compiles the two projects managed with bras that I have easily.
While writing the doc, I found out that it is rather illogical to not
allow more than one command for a target. This is what I am going to implement
now.

Consequently, the doc is in a transient state, i.e. it neither fully
describes this version, nor the coming one.