Until now, sourcing of brasfiles in directories were @-dependencies lead to happened as soon as such a dependency was seen in a rule.
Now this is delayed until the dependency is really considered as a target,
resulting in considerable speedup, in particular, if bras is called in a
subdirectory, where the sibling directories are of no interest.

To be able to compare with the old behaviour, option -es for `early sourcing'
was introduced.