Several major changes.
o In a prerelease, the DependsFile-rule was called CDepends.

o The whole method of using pattern-rules was revised and changed. Basically
  the order in which they were checked was changed. Also MatchDep* was
  introduced to help finding default commands. The result of MatchDep* is
  made available to the command to be executed in global variable
  patternTrigger. Finding default rules now works recursively.

o Option -n does *really* nothing now.

o The old behaviour of -n was moved to -N.

o Options -s and -ss allow silent and super-silent operation.

o The `include'-command now sets global variable brasLastInclude to the
  file it is going to source.

All this required the change of several tests.