A major change with regard to the handling of environment variables. The previous idea of copying them into global variables while controlling with -e when this has to happen was inspired by make but does not work well with bras, in particular because globals may be set in Brasfiles which are only source on demand of an @-dependency.
Instead of copying environment variables into globals, there is now a
builtin command getenv which may be used like
  getenv prefix /my/special/dir
It sets prefix to the given value, except if the environement contains a
value for prefix. In that case, prefix is set to the value from the
environment. As before, the environment can be extended on the commandline
with a syntax like var='new value'.

I also fiddled with evalCmds.tcl. It was broken with respect to handling
commands to be auto-loaded. The new version should handle auto-loaded
functions correctly, however I did not yet test it. Also the output in case
of errors was changed slightly.