This is called 0.99.0 and can be understood as pre-1.0.0.
Major changes have happened:

-- Bras is now be a loadable package. The main script `bras' only serves
as a front-end with the same functionality as before. But with
[package require bras] [namespace import ::bras::*] the whole
machinery can now be made available in just any script, even within
TclX or a wish. Everthing is (should be) now in namespace ::bras.

-- There are no more different rule types. Only one type of rule
exists. It has the form

   Make targets boolean-expression command

where a typical example is

   Make bla.o {[older bla.o bla.c]} {
     cc -o $target [lindex $deps 0]
   }

which can be translated into English like "make bla.o if bla.o is
older than bla.c with the given command".

-- To be compatible with previous versions, Newer, Exist and Always
still exist, but are mappend internally into calls to `Make'.


## INCOMPATIBILITIES

-- Since the `boolean-expression'-part can call just about any
function, it can in particular compute dependencies on the
fly. Consequently `Defdexpand', a proc previously used to define
dynamic dependency generators no longer exists. The best way to
translate this to the new version for a dependency generator called
`bla' is probably to write a `proc bla' which returns the dependencies
and then use `Make this [older this [bla]] { }'.

-- Since there are no more rule-types, `Defrule' does no longer
exist. Instead, a correlating predicate to be used in
`boolean-expression' must be written.

-- Commands GetDeps, GetCmd, ClearDeps, ClearCmd which allowed direct
access to the internal rule database are removed. I don't feel they
are really useful. (Prove me wrong and I'll put them in again.)

## STATUS
-- Dispite the incompatibilities, two of my projects (one using
`searchpath') were built with only a slight change in one of them
caused by the removal of Defdexpand.

-- The docs are not yet updated.