o Added predicate varchanged. With the help of a cache file, it is able to check if a parameter stored in an init-file was changed since the last invocation. This allows to trigger computations with changes of individual values set in a file.
o Added function forget. It tells bras to forget that it has
considered a given target already. This is most useful in
GUI-applications which want to reconsider targets, after the user made
changes to parameters.

o Also for GUI-applications, configuration flag -ve can be reset.

o Again for GUI-applications, all `puts' calls are now routed through
a central function called `::bras::report' which also gets a parameter
telling the type of output to produce. By overwriting this function,
an applications is able to redirect output into a widget.

o Error handling and exiting was streamlined. Bras does no longer
contain a call to `exit'. Serious errors now throw exceptions. To mark
the end of the information relevant to the user on the stack, the
string "---SNIP---" is entered into the error traceback.

o This lead to version 0.99.5.