This will be version 1.3.0.
Added a pattern rule to install files (install.rule). A target
description like /usr/bin/bla/0755 is made dependent on the file bla
in the local directory. If the local file is newer it is copied over
the target, /usr/bin/bla, and then the permissions are set as
requested by the suffixed numbers.

This needed a new proc, namely install, which does the copying and
permission fiddling.

Docs updated accordingly. And example is also available.