This is going to be 1.4.1. I will not be published since it does not contain new features.
The file pkgIndex.tcl is no longer autogenerated but prepared by
hand. It contains now the version number and allows to get rid of
the file .version.

Finally removed subdirectories Test1, Test2 which were not used for
quite some time now.

Removed the non-feature of bras being independent of the package. This
was only a half-solution because it did not take rule-files into
account.